# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API-docs of backend server of Harmonica team.",
        "title": "Harmonium backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "https://harmoniums.ru",
    "basePath": "api/v1",
    "paths": {
...
        "/users/{user_id}": {
            "post": {
                "description": "Update user by description and user id.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "type": "string",
                        "default": "session-token=",
                        "description": "session-token",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User information in json",
                        "name": "user",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "User avatar",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.PinPageResponse"
                        }
                    },
                    "400": {
                        "description": "Possible code responses: 3, 4, 5, 12, 13, 18",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Possible code responses: 2.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Possible code responses: 14.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Possible code responses: 6, 11.",
                        "schema": {
                            "$ref": "#/definitions/errs.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entity.PinAuthor": {
            "description": "User-author information with user id, nickname and avatar",
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "entity.PinPageResponse": {
            "description": "Full pin information",
            "type": "object",
            "properties": {
                "allow_comments": {
                    "type": "boolean"
                },
                "author": {
                    "$ref": "#/definitions/entity.PinAuthor"
                },
                "click_url": {
                    "type": "string"
                },
                "content_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_owner": {
                    "type": "boolean"
                },
                "likes_count": {
                    "type": "integer"
                },
                "pin_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "errs.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}